BSPQNXIMX8X-29 src: utils: usbcmgr: Add a utilty for TPS65981 status/control
The USB-C controller chip can run in host/DFP mode entirely from the
SPI flash configuration.  We are not sure if configuration changes
will be needed if/when QNX supports device/UFP on the USB_OTG2
interface that the chip is connected to.  But this provides a
framework for chip setup over I2C.  It also provides status
information which is at least handy to verify that the SPI flash
is programmed correctly.

Signed-off-by: Steve Schefter <steve@scheftech.com>