BSPQNXIMX8X-23 src: Add scripts for starting USB
The QNX USB stack does not support OTG.  Rather, you need to start
up the stack with the DLLs that provide either host or device
support, not both.  In order to provide more flexibility, the
USB startup is commented out of the images/*.build files
allowing them to be started on the command line.  Since the
startup involves longish commands, these scripts are provided.

Signed-off-by: Steve Schefter <steve@scheftech.com>