BSPQNXIMX8X-31 src: hardware: Add a wfdcfg library for PEB-LVDS-01
That is, the TDA19988 LVDS-to-HDMI controller, which is the only
programmable part on PEB-LVDS-01.

By default, the BSP is configured for an LCD-018 on each of the
LVDS intefaces.  See usage notes for configuration changes to
enable PEB-LVDS-01 on one or both of the interfaces.

This wfdcfg library is used by the graphics driver to report the
timing and resolution (modes) supported by the monitor as learned
by reading the EDID information.

Signed-off-by: Steve Schefter <steve@scheftech.com>